Some cleanup of startup code. electron-main has gotten too large and hard to test.
